# Conflicts:
#	api/http/handler/kubernetes/handler.go

+++ app/docker/views/secrets/secretsController.js
@@ -1,10 +1,15 @@
+import { confirmDeletionAsync } from 'Portainer/services/modal.service/confirm';
 angular.module('portainer.docker').controller('SecretsController', [
   '$scope',
   '$state',
   'SecretService',
   'Notifications',
   function ($scope, $state, SecretService, Notifications) {
-    $scope.removeAction = function (selectedItems) {
+    $scope.removeAction = async function (selectedItems) {
+      const confirmed = await confirmDeletionAsync('Do you want to remove the selected secret(s)?');
+      if (!confirmed) {
+        return null;
+      }
       var actionCount = selectedItems.length;
       angular.forEach(selectedItems, function (secret) {
         SecretService.remove(secret.Id)
